+++ storage-reverse-image-search/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 fix - bump to nodejs20 runtime in functions and run npm audit fix
 
+fix - if backfill is disabled, create index on first image upload
+
 ## Version 0.1.8
 
 fix - update default Gemini models